+++ core/interfaces.go
@@ -44,6 +44,8 @@ type Services interface {
 	ReportGroupAsAbuse(clientID string, current *model.User, group *model.Group, comment string) error
 
 	GetGroup(clientID string, current *model.User, id string) (*model.Group, error)
+	GetGroupsByGroupIDs(groupIDs []string) ([]model.Group, error)
+
 	GetGroupStats(clientID string, id string) (*model.GroupStats, error)
 
 	ApplyMembershipApproval(clientID string, current *model.User, membershipID string, approve bool, rejectReason string) error
@@ -56,6 +58,8 @@ type Services interface {
 	DeleteEvent(clientID string, current *model.User, eventID string, groupID string) error
 	GetEventUserIDs(eventID string) ([]string, error)
 	GetGroupMembershipsStatusAndGroupTitle(userID string) ([]model.GetGroupMembershipsResponse, error)
+	GetGroupMembershipsByGroupID(groupID string) ([]string, error)
+
 	GetGroupsEvents(eventIDs []string) ([]model.GetGroupsEvents, error)
 
 	GetPosts(clientID string, current *model.User, filter model.PostsFilter, filterPrivatePostsValue *bool, filterByToMembers bool) ([]model.Post, error)
@@ -211,9 +215,15 @@ func (s *servicesImpl) GetEventUserIDs(eventID string) ([]string, error) {
 func (s *servicesImpl) GetGroupMembershipsStatusAndGroupTitle(userID string) ([]model.GetGroupMembershipsResponse, error) {
 	return s.app.findGroupMembershipsStatusAndGroupsTitle(userID)
 }
+func (s *servicesImpl) GetGroupMembershipsByGroupID(groupID string) ([]string, error) {
+	return s.app.findGroupMembershipsByGroupID(groupID)
+}
 
 func (s *servicesImpl) GetGroupsEvents(eventIDs []string) ([]model.GetGroupsEvents, error) {
 	return s.app.findGroupsEvents(eventIDs)
+}
+func (s *servicesImpl) GetGroupsByGroupIDs(groupIDs []string) ([]model.Group, error) {
+	return s.app.findGroupsByGroupIDs(groupIDs)
 }
 
 func (s *servicesImpl) GetPosts(clientID string, current *model.User, filter model.PostsFilter, filterPrivatePostsValue *bool, filterByToMembers bool) ([]model.Post, error) {
@@ -419,6 +429,7 @@ type Storage interface {
 	FindGroup(context storage.TransactionContext, clientID string, groupID string, userID *string) (*model.Group, error)
 	FindGroupByTitle(clientID string, title string) (*model.Group, error)
 	FindGroups(clientID string, userID *string, filter model.GroupsFilter) ([]model.Group, error)
+	FindGroupsByGroupIDs(groupIDs []string) ([]model.Group, error)
 	FindUserGroups(clientID string, userID string, filter model.GroupsFilter) ([]model.Group, error)
 	FindUserGroupsCount(clientID string, userID string) (*int64, error)
 	DeleteUsersByAccountsIDs(log *logs.Logger, context storage.TransactionContext, accountsIDs []string) error
@@ -431,6 +442,8 @@ type Storage interface {
 
 	FindEventUserIDs(context storage.TransactionContext, eventID string) ([]string, error)
 	FindGroupMembershipStatusAndGroupTitle(context storage.TransactionContext, userID string) ([]model.GetGroupMembershipsResponse, error)
+	FindGroupMembershipByGroupID(context storage.TransactionContext, groupID string) ([]string, error)
+
 	FindGroupsEvents(context storage.TransactionContext, eventIDs []string) ([]model.GetGroupsEvents, error)
 
 	ReportGroupAsAbuse(clientID string, userID string, group *model.Group) error