+++ core/services_events.go
@@ -39,8 +39,15 @@ func (app *Application) findGroupMembershipsStatusAndGroupsTitle(userID string) 
 	return app.storage.FindGroupMembershipStatusAndGroupTitle(nil, userID)
 }
 
+func (app *Application) findGroupMembershipsByGroupID(groupID string) ([]string, error) {
+	return app.storage.FindGroupMembershipByGroupID(nil, groupID)
+}
 func (app *Application) findGroupsEvents(eventIDs []string) ([]model.GetGroupsEvents, error) {
 	return app.storage.FindGroupsEvents(nil, eventIDs)
+}
+
+func (app *Application) findGroupsByGroupIDs(groupIDs []string) ([]model.Group, error) {
+	return app.storage.FindGroupsByGroupIDs(groupIDs)
 }
 
 func (app *Application) getEvents(clientID string, current *model.User, groupID string, filterByToMembers bool) ([]model.Event, error) {